+++ wemake_python_styleguide/options/validation.py
@@ -88,11 +88,13 @@ class _ValidatedOptions(object):
     max_asserts: int = attr.ib(validator=[_min_max(min=1)])
     max_access_level: int = attr.ib(validator=[_min_max(min=1)])
     max_attributes: int = attr.ib(validator=[_min_max(min=1)])
+    max_raises: int = attr.ib(validator=[_min_max(min=1)])
     max_cognitive_score: int = attr.ib(validator=[_min_max(min=1)])
     max_cognitive_average: int = attr.ib(validator=[_min_max(min=1)])
     max_call_level: int = attr.ib(validator=[_min_max(min=1)])
     max_annotation_complexity: int = attr.ib(validator=[_min_max(min=2)])
     max_import_from_members: int = attr.ib(validator=[_min_max(min=1)])
+    max_tuple_unpack_length: int = attr.ib(validator=[_min_max(min=1)])
 
 
 def validate_options(options: ConfigurationOptions) -> _ValidatedOptions: