+++ wemake_python_styleguide/violations/base.py
@@ -35,14 +35,14 @@ Conventions
 Deprecating a violation
 ~~~~~~~~~~~~~~~~~~~~~~~
 
-When you want to mark some violation as depracated,
+When you want to mark some violation as deprecated,
 then assign ``deprecated`` boolean flag to it:
 
 .. code:: python
 
   @final
   class SomeViolation(ASTViolation):
-      depracated = True
+      deprecated = True
 
 Reference
 ~~~~~~~~~
@@ -68,11 +68,9 @@ ErrorNode = Union[
 ErrorCallback = Callable[['BaseViolation'], None]
 
 
+@enum.unique
 class ViolationPostfixes(enum.Enum):
     """String values of postfixes used for violation baselines."""
-
-    #: This field is required for `mypy` plugin that types field values.
-    value: str  # noqa: WPS110
 
     bigger_than = ' > {0}'
     less_than = ' < {0}'
@@ -89,7 +87,7 @@ class BaseViolation(object, metaclass=abc.ABCMeta):
 
     Attributes:
         error_template: message that will be shown to user after formatting.
-        code: violation unique number. Used to identify the violation.
+        code: unique violation number. Used to identify the violation.
         previous_codes: just a documentation thing to track changes in time.
         deprecated: indicates that this violation will be removed soon.
         postfix_template: indicates message that we show at the very end.